refactor(app-shell): 12 mutation を useDashboardMutations に集約 (refs #152)
AppShell.tsx 責務分離 (issue #152) の Chunk 3 (本命)。挙動は変えない。

- `useDashboardMutations`: toggleDone / updateBody / updateDependency /
  updateCategory / reorder / createTaskWithAi / createEvent /
  updateEventProject / updateEvent / deleteEvent / createProject /
  updateProject / deleteProject / deleteTask の 14 操作 + AI trigger 派生
  callback 2 個 (triggerDecomposeWithOptimistic /
  triggerResplitWithOptimistic) を 1 hook に集約。
- 各 mutation は optimistic update + onError でロールバック + onSettled
  invalidate の既存パターンをそのまま移植。action_log 記録 (TASK_COMPLETED /
  TASK_DEPENDENCY_SET / _CLEARED / TASK_CATEGORY_CHANGED / TASK_REORDERED) も
  同位置で実行。
- AppShell では分割代入で受け取り、JSX に直接渡す形に変更。
  inline arrow function による props 用 wrapper も削減。
- 不要になった import (ACTION_TYPES / log / triggerDecompose 等) を AppShell
  から除去。

AppShell.tsx は 770 → 378 行 (約 -50%)。
